MAINTAINERS: Update Grant's email address and maintainership
I've taken a full time position with Linaro and so I'll be using my
Linaro email address from this point on. It has also been many years
since I've touched any of the Xilinx related code so mark those items as
unmaintained.

In addition, Mark Brown is taking the lead on SPI maintainership now, so
I've reversed the order of our names for that entry.

Signed-off-by: Grant Likely <grant.likely@linaro.org>
Cc: Mark Brown <broonie@opensource.wolfsonmicro.com>
Cc: Linus Walleij <linus.walleij@linaro.org>
Cc: Rob Herring <rob.herring@calxeda.com>